refactor application/defaults.rb to concern
